agent/pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "AI agent for CPE project"
authors = [
    {name = "cpe",email = "cpe@fpt.com"}
]
readme = "README.md"
requires-python = "3.10.11"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "langchain (>=0.3.23,<0.4.0)",
    "langchain-openai (>=0.3.12,<0.4.0)",
    "langgraph (>=0.3.27,<0.4.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "langchain-chroma (>=0.2.2,<0.3.0)",
    "langchain-community (>=0.3.21,<0.4.0)",
    "azure-storage-blob (>=12.25.1,<13.0.0)",
    "langchain-postgres (>=0.0.14,<0.0.15)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "types-pyyaml (>=6.0.12.20250402,<7.0.0.0)",
    "types-psycopg2 (>=2.9.21.20250318,<3.0.0.0)",
    "pgvector (==0.3.6)",
    "pyodbc (>=5.2.0,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "types-python-jose (>=3.4.0.20250224,<4.0.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "types-passlib (>=1.7.7.20250408,<2.0.0.0)",
    "bcrypt (==4.0.1)",
]

[tool.poetry]
packages = [{include = "ai_agent", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
ipykernel = "^6.29.5"
alembic = "^1.15.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
